added the people migration and model.
I am currently only building out the basic tests 'is valid with valid arguments' and 'in not valid without email attribute'. I feel like there should be more here but I am not sure what the best tests to run on a simple object like this is. I have reached out to Mimi for her direction and in the meantime I am going to continue with the peoples controller spec and route.

Next:
 - peoples controller spec
 - peoples route
 - peoples controller